update and rename entities files
