+++ src/products/entities/presentation.entity.ts
@@ -1,10 +1,10 @@
 import { Column, Entity, OneToMany } from 'typeorm';
-import { ProductPresentation } from './product.presentation.entity';
+import { ProductPresentation } from './product-presentation.entity';
 import { BaseModel } from 'src/utils/entity';
 
 @Entity('presentation')
 export class Presentation extends BaseModel {
-  @Column({ type: 'varchar', name: 'name' })
+  @Column({ type: 'character varying', name: 'name' })
   name: string;
 
   @Column({ type: 'text', name: 'description' })
@@ -13,7 +13,7 @@ export class Presentation extends BaseModel {
   @Column({ type: 'int', name: 'quantity' })
   quantity: number;
 
-  @Column({ type: 'varchar', name: 'measurement_unit' })
+  @Column({ type: 'character varying', name: 'measurement_unit' })
   measurementUnit: string;
 
   @OneToMany(
